requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
cssutils==2.3.0
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.3
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-imagekit==4.0.2
django-inlinecss==0.3.0
django-nested-admin==3.3.3
django-richtextfield==1.6.1
django-social-share==2.2.1
filelock==3.0.12
future==0.18.2
idna==2.10
kombu==5.1.0
oauthlib==3.1.0
packaging==20.9
pilkit==2.0
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pynliner==0.8.0
pyparsing==2.4.7
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
vine==5.0.0
virtualenv==20.4.7
wcwidth==0.2.5
